cpu/percpu: store a safe reference to the HV doorbell page
Mutable references to the HV doorbell page are never given, as its
data is fully atomic. Thus, there is no need to use raw pointers.

Since we need a pointer to the location of the safe reference itself
in the PerCpu for #HV servicing, use a OnceCell, which gives out
references to the inner T under the assumption that the type can only
be set once, which suits our use. With a &T (in this case a &&'static
HVDoorbell), we can obtain a pointer to the inner value.

+++ kernel/src/cpu/idt/svsm.rs
@@ -110,7 +110,7 @@ pub fn idt_init() {
     // that uses the value is not type safe in any case, so enforcing type
     // safety on the pointer would offer no meaningful value.
     unsafe {
-        HV_DOORBELL_ADDR = this_cpu().hv_doorbell_addr();
+        HV_DOORBELL_ADDR = this_cpu().hv_doorbell_addr() as usize;
     };
 }
